feat(hub-common): remove unnecessary import line moves
affects: @esri/hub-common
